Moved code out of __init__
